feat-chat-stream: stream 구조 확인
