+++ ganoverflow-next/package.json
@@ -20,6 +20,7 @@
     "axios": "^1.4.0",
     "eslint": "8.42.0",
     "eslint-config-next": "13.4.4",
+    "eventsource-parser": "^1.0.0",
     "hangul-js": "^0.2.6",
     "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.0",
