+++ ganoverflow-next/src/styles/globals.css
@@ -94,7 +94,7 @@ body,
     "Segoe UI Symbol";
   -webkit-font-smoothing: antialiased;
   text-rendering: optimizelegibility;
-  color: rgb(248, 248, 248);
+  color: rgb(118, 118, 118);
   background-color: #1c2128;
 }
